Use the QGis EarthEngine plugin
The QGis EarthEngine plugin provides all necessary features to create a QGis layer that contains the Earth Engine map. Additionally it enables to save/open the EE layer in a QGis project. Instead of redeveloping these features, this plugin focuses on the GUI to create and configure the Sentinel-2 images rendering.